requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
rdkit==2023.3.3
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7